feat(sdk): errors added
